pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "modern-elt-pipeline"
version = "0.1.0"
description = "A local modern ELT pipeline using Python, Postgres, dbt, Dagster, and data quality checks."
requires-python = ">=3.10"
dependencies = [
  "dagster==1.10.8",
  "dagster-webserver==1.10.8",
  "dbt-core==1.9.4",
  "dbt-postgres==1.9.0",
  "pandas==2.1.4",
  "psycopg2-binary==2.9.10",
  "SQLAlchemy==2.0.40",
  "python-dotenv==1.1.0",
  "pydantic==2.11.3",
  "pydantic-settings==2.8.1"
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.5",
  "ruff==0.11.6"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
